feat: send messages to unconnected accounts
